+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Deprecated
 
 * `RealmResults.removeChangeListeners()`. Use `RealmResults.removeAllChangeListeners()` instead.
+* `RealmObject.removeChangeListeners()`. Use `RealmObject.removeAllChangeListeners()` instead.
 
 ### Enhancements